move FormBox elements in a single file
