--- packages/pn-pa-webapp/src/components/FormBox/FormBoxTitle.tsx
@@ -1,11 +1,0 @@
-import { Typography } from '@mui/material';
-
-type Props = {
-  text: string;
-};
-
-export const FormBoxTitle = ({ text }: Props) => (
-  <Typography variant="sidenav" fontWeight={600} fontSize={'16px'}>
-    {text}
-  </Typography>
-);
